+++ .github/workflows/classify_workflow_errors.py
@@ -23,7 +23,7 @@ from datetime import datetime, timezone
 # Add project root to path
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
-from opensearch_upload.error_classification import (
+from error_classification import (
     Config,
     ErrorClassifier,
     ErrorContext,
@@ -401,141 +401,6 @@ def extract_errors_from_workflow() -> List[ErrorContext]:
             print(f"  ℹ️  Created generic error for {job_name}")
 
     return errors
-
-
-def classify_and_annotate_workflow_errors():
-    """
-    Validate classifications for consistency and quality.
-
-    Returns list of validation issues found.
-    """
-    issues = []
-
-    # Build error hash to classification mapping
-    from opensearch_upload.error_classification.deduplicator import ErrorDeduplicator
-    dedup = ErrorDeduplicator(None, None)
-
-    hash_to_classifications = {}
-    for i, (classification, error) in enumerate(zip(classifications, errors)):
-        error_hash = dedup.compute_error_hash(error.error_text)
-
-        if error_hash not in hash_to_classifications:
-            hash_to_classifications[error_hash] = []
-
-        hash_to_classifications[error_hash].append({
-            'index': i + 1,
-            'job': error.job_name,
-            'category': classification.primary_category,
-            'confidence': classification.confidence_score,
-            'error_text': error.error_text[:100]
-        })
-
-    # Check for duplicate errors with different classifications
-    for error_hash, group in hash_to_classifications.items():
-        if len(group) > 1:
-            # Multiple errors with same hash
-            categories = set(item['category'] for item in group)
-
-            if len(categories) > 1:
-                # Same error, different classifications - VALIDATION FAILURE
-                jobs = ", ".join([item['job'] for item in group])
-                cats = ", ".join([f"{item['category']}({item['confidence']:.0%})" for item in group])
-
-                issues.append(
-                    f"Duplicate errors classified differently:\n"
-                    f"      Jobs: {jobs}\n"
-                    f"      Classifications: {cats}\n"
-                    f"      Error: {group[0]['error_text']}..."
-                )
-            else:
-                # Same error, same classification - GOOD
-                print(f"  ✅ Deduplication validated: {len(group)} jobs with identical error -> {group[0]['category']}")
-
-    # Enhanced Confidence Validation
-    confidences = [c.confidence_score for c in classifications]
-    avg_confidence = sum(confidences) / len(confidences) if confidences else 0
-
-    print(f"  📊 Confidence distribution:")
-    print(f"     Average: {avg_confidence:.1%}")
-    print(f"     Min: {min(confidences):.1%}, Max: {max(confidences):.1%}")
-
-    # Check 1: Very low confidence classifications
-    low_confidence = [
-        (i + 1, c.primary_category, c.confidence_score, errors[i].job_name)
-        for i, c in enumerate(classifications)
-        if c.confidence_score < 0.6
-    ]
-
-    if low_confidence:
-        for idx, cat, conf, job in low_confidence:
-            issues.append(
-                f"⚠️  Low confidence (#{idx}): {job} -> {cat} ({conf:.0%})"
-            )
-
-    # Check 2: Average confidence too low (suggests systematic issue)
-    if avg_confidence < 0.65:
-        issues.append(
-            f"⚠️  Average confidence very low ({avg_confidence:.1%}) - may indicate extraction issues"
-        )
-
-    # Check 3: High variance in confidence for duplicate errors
-    for error_hash, group in hash_to_classifications.items():
-        if len(group) > 1:
-            confidences_in_group = [item['confidence'] for item in group]
-            confidence_variance = max(confidences_in_group) - min(confidences_in_group)
-
-            if confidence_variance > 0.15:  # More than 15% difference
-                issues.append(
-                    f"⚠️  High confidence variance for duplicate errors:\n"
-                    f"      Jobs: {', '.join([item['job'] for item in group])}\n"
-                    f"      Confidences: {', '.join([f'{c:.0%}' for c in confidences_in_group])}"
-                )
-
-    # Pattern validation: Check if error text matches category expectations
-    pattern_mismatches = []
-    for i, (classification, error) in enumerate(zip(classifications, errors)):
-        error_text = error.error_text.lower()
-        category = classification.primary_category
-
-        # Define expected patterns for each category
-        expected_patterns = {
-            'infrastructure_error': ['importerror', 'modulenotfounderror', 'error collecting', 'cannot import'],
-            'timeout': ['timeout', 'timed out', 'exceeded time limit', 'deadline exceeded'],
-            'network_error': ['connection', 'network', 'dns', 'refused', 'unreachable'],
-            'dependency_error': ['could not find', 'no matching distribution', 'version conflict'],
-            'assertion_failure': ['assertionerror', 'assert ', 'expected', 'actual'],
-        }
-
-        # Check if category matches patterns
-        if category in expected_patterns:
-            patterns = expected_patterns[category]
-            if not any(pattern in error_text for pattern in patterns):
-                # Category doesn't match error text patterns
-                pattern_mismatches.append(
-                    f"Pattern mismatch (#{i + 1}): {error.job_name} classified as {category} "
-                    f"but error text doesn't match expected patterns"
-                )
-
-    # Only report pattern mismatches if confidence is also low
-    for i, (classification, error) in enumerate(zip(classifications, errors)):
-        if classification.confidence_score < 0.7:
-            error_text = error.error_text.lower()
-            category = classification.primary_category
-
-            expected_patterns = {
-                'infrastructure_error': ['importerror', 'modulenotfounderror', 'error collecting'],
-                'timeout': ['timeout', 'timed out', 'exceeded'],
-                'network_error': ['connection', 'network', 'dns'],
-            }
-
-            if category in expected_patterns:
-                if not any(pattern in error_text for pattern in expected_patterns[category]):
-                    issues.append(
-                        f"Low confidence + pattern mismatch (#{i + 1}): "
-                        f"{error.job_name} -> {category} ({classification.confidence_score:.0%})"
-                    )
-
-    return issues
 
 
 def classify_and_annotate_workflow_errors():
